event page added and some bug fixed
