+++ src/components/Footer.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import '../css/Footer.css'
-function Footer(props){
+function FooterLanding(props){
    return ( 
       <div class="parent-element">
          <div className={`footer ${props.bgColor}`}>
@@ -16,11 +16,11 @@ function Footer(props){
                   </div>
                   <div class="links">
                         <ul class="footer-links flex-container">
-                           <li><a href="https://www.linkedin.com/company/geeksforgeeks-kiit"><i class="fab fa-linkedin"></i></a></li>
-                           <li><a href="https://youtube.com/channel/UC281oAcZl0asjjuvmnXh2PQ"><i class="fab fa-youtube"></i></a></li>
-                           <li><a href="https://github.com/GFG-CLUB-KIIT"><i class="fab fa-github"></i></a></li>
-                           <li><a href="#"><i class="fab fa-twitter"></i></a></li>
-                           <li><a href="https://instagram.com/gfg_kiit_club?igshid=rnewwd09app3"><i class="fab fa-instagram"></i></a></li>
+                           <li><a href="https://www.linkedin.com/company/geeksforgeeks-kiit" target="_blank"><i class="fab fa-linkedin"></i></a></li>
+                           <li><a href="https://youtube.com/channel/UC281oAcZl0asjjuvmnXh2PQ" target="_blank"><i class="fab fa-youtube"></i></a></li>
+                           <li><a href="https://github.com/GFG-CLUB-KIIT"  target="_blank"><i class="fab fa-github"></i></a></li>
+                           <li><a href="https://twitter.com/ClubGfg?s=08" target="_blank"><i class="fab fa-twitter" ></i></a></li>
+                           <li><a href="https://instagram.com/gfg_kiit_club?igshid=rnewwd09app3" target="_blank"><i class="fab fa-instagram"></i></a></li>
                         </ul>
                   </div>
                   <div class="footer-info">
@@ -32,24 +32,24 @@ function Footer(props){
                            <p>Engineering, KIIT</p>
                            <p>University, Bhubaneswar</p>
                            <p>751024</p>
-                           <p>Email:  gfgkiit@gmail.com</p>
-                           <p>Phone: +91 00000 00000</p>
-                           <p>(Veer Sharma)</p>
+                           <p>Email: gfgkiit@gmail.com</p>
+                           <p>Phone: +91 87655 16718</p>
+                           <p>(Taran)</p>
+                           <br></br>
                            </div>
                         <div>
-                        <div class="description2">
-                           <p class = "community strong">Community</p>
-                           <br></br>
-                           <p>asdadsddada</p>
-                           <p>GFG KIIT</p>
-                           <p>asdas</p>
-                        </div>
-                        <div class="description3">
-                           <p class = "guideline strong">General Guidelines</p>
-                           <br></br>
-                           <p>Code of</p>
-                           <p>Conduct</p>
-                           <p></p>
+                           <div class="description3">
+                              <p class = "guideline strong">General Guidelines</p>
+                              <br></br>
+                              <p>Code of Conduct</p>    
+                              <br></br>        
+                              <br></br>        
+                           </div>
+                           <div class="description2">
+                              <p class = "community strong">Community</p>
+                              <br></br>
+                              <p>GFG KIIT</p>
+                              <br></br>
                         </div>
                      </div>
                   </div>
@@ -61,4 +61,4 @@ function Footer(props){
    )
 }
 
-export default Footer
+export default FooterLanding
