+++ src/components/landing-scroll-banding/LandingScrollBranding.js
@@ -1,0 +1,91 @@
+import React from 'react'
+import './landing-scroll-branding.css'
+
+export default function LandingScrollBranding() {
+   return (
+      <div>
+         <section class="landing-scroll-branding">
+            <div class="scroll designers org">
+               <div>
+                  <span class="filled">Designers</span> <span>Developers</span>
+                  <span>Programmers</span> <span>Creators</span>
+                  <span>Enthusiasts</span> <span class="filled">Designers</span>
+                  <span>Developers</span> <span>Programmers</span>
+                  <span>Creators</span> <span>Enthusiasts</span>
+               </div>
+               <div>
+                  <span class="filled">Designers</span> <span>Developers</span>
+                  <span>Programmers</span> <span>Creators</span>
+                  <span>Enthusiasts</span> <span class="filled">Designers</span>
+                  <span>Developers</span> <span>Programmers</span>
+                  <span>Creators</span> <span>Enthusiasts</span>
+               </div>
+            </div>
+            <div class="scroll developers alt">
+               <div>
+                  <span>Designers</span> <span class="filled">Developers</span>
+                  <span>Programmers</span> <span>Creators</span>
+                  <span>Enthusiasts</span> <span>Designers</span>
+                  <span class="filled">Developers</span> <span>Programmers</span>
+                  <span>Creators</span> <span>Enthusiasts</span>
+               </div>
+               <div>
+                  <span>Designers</span> <span class="filled">Developers</span>
+                  <span>Programmers</span> <span>Creators</span>
+                  <span>Enthusiasts</span> <span>Designers</span>
+                  <span class="filled">Developers</span> <span>Programmers</span>
+                  <span>Creators</span> <span>Enthusiasts</span>
+               </div>
+            </div>
+            <div class="scroll programmers org">
+               <div>
+                  <span>Designers</span> <span>Developers</span>
+                  <span class="filled">Programmers</span> <span>Creators</span>
+                  <span>Enthusiasts</span> <span>Designers</span>
+                  <span>Developers</span> <span class="filled">Programmers</span>
+                  <span>Creators</span> <span>Enthusiasts</span>
+               </div>
+               <div>
+                  <span>Designers</span> <span>Developers</span>
+                  <span class="filled">Programmers</span> <span>Creators</span>
+                  <span>Enthusiasts</span> <span>Designers</span>
+                  <span>Developers</span> <span class="filled">Programmers</span>
+                  <span>Creators</span> <span>Enthusiasts</span>
+               </div>
+            </div>
+            <div class="scroll creators alt">
+               <div>
+                  <span>Designers</span> <span>Developers</span>
+                  <span>Programmers</span> <span class="filled">Creators</span>
+                  <span>Enthusiasts</span> <span>Designers</span>
+                  <span>Developers</span> <span>Programmers</span>
+                  <span class="filled">Creators</span> <span>Enthusiasts</span>
+               </div>
+               <div>
+                  <span>Designers</span> <span>Developers</span>
+                  <span>Programmers</span> <span class="filled">Creators</span>
+                  <span>Enthusiasts</span> <span>Designers</span>
+                  <span>Developers</span> <span>Programmers</span>
+                  <span class="filled">Creators</span> <span>Enthusiasts</span>
+               </div>
+            </div>
+            <div class="scroll enthusiasts org">
+               <div>
+                  <span>Designers</span> <span>Developers</span>
+                  <span>Programmers</span> <span>Creators</span>
+                  <span class="filled">Enthusiasts</span> <span>Designers</span>
+                  <span>Developers</span> <span>Programmers</span>
+                  <span>Creators</span> <span class="filled">Enthusiasts</span>
+               </div>
+               <div>
+                  <span>Designers</span> <span>Developers</span>
+                  <span>Programmers</span> <span>Creators</span>
+                  <span class="filled">Enthusiasts</span> <span>Designers</span>
+                  <span>Developers</span> <span>Programmers</span>
+                  <span>Creators</span> <span class="filled">Enthusiasts</span>
+               </div>
+            </div>
+         </section>
+      </div>
+   )
+}
