+++ src/components/landing-scroll-banding/landing-scroll-branding.css
@@ -1,0 +1,80 @@
+.landing-scroll-branding {
+   position: relative;
+   width: 100%;
+   height: 100%;
+   overflow: hidden;
+   font-family: "Montserrat";
+}
+.landing-scroll-branding .scroll {
+   width: 100%;
+   display: flex;
+}
+.landing-scroll-branding .scroll {
+   animation-play-state: paused;
+}
+
+.landing-scroll-branding .scroll div {
+   color: rgba(0, 0, 0, 0);
+   font-size: 3em;
+   white-space: nowrap;
+   padding: 10px;
+   font-weight: 800;
+   text-transform: none;
+   -webkit-text-stroke-width: 1px;
+   -webkit-text-stroke-color: rgba(0, 0, 0, 0.75);
+}
+.landing-scroll-branding .scroll .filled {
+   color: rgba(0, 0, 0, 0.75);
+}
+.landing-scroll-branding .org div {
+   animation: animate 80s linear infinite;
+   animation-delay: -80s;
+}
+.landing-scroll-branding .org div:nth-child(2) {
+   animation: animate2 80s linear infinite;
+   animation-delay: -40s;
+}
+
+.landing-scroll-branding .alt div {
+   animation: animate-alt 80s linear infinite;
+   animation-delay: -80s;
+}
+
+.landing-scroll-branding .alt div:nth-child(2) {
+   animation: animate-alt2 80s linear infinite;
+   animation-delay: -40s;
+}
+
+@keyframes animate {
+   0% {
+      transform: translateX(100%);
+   }
+   100% {
+      transform: translateX(-100%);
+   }
+}
+@keyframes animate2 {
+   0% {
+      transform: translateX(0%);
+   }
+   100% {
+      transform: translateX(-200%);
+   }
+}
+
+@keyframes animate-alt {
+   0% {
+      transform: translateX(-100%);
+   }
+   100% {
+      transform: translateX(100%);
+   }
+}
+@keyframes animate-alt2 {
+   0% {
+      transform: translateX(-200%);
+   }
+   100% {
+      transform: translateX(0%);
+   }
+}
