--- src/css/AboutUs.css
@@ -1,192 +1,0 @@
-body{
-    margin: 0px;
-    padding: 0px;
-    font-family: pt sans !important;
-    overflow-x: hidden;
-}
-#aboutUs_title{
-    width: 400px;
-    float: left!important;
-    color: white;
-    margin-left: 70%;
-    margin-top: -300px;
-    position:absolute;
-}
-#aboutUs_title_right{
-    font-size: 40px;
-    font-weight: 500;
-    font-family: sans-serif;
-    letter-spacing: 1.5px;
-}
-#aboutUs_body{
-    padding: 0px 5%;
-}
-#cm_title{
-    width: 100%;
-    margin-top: 50px;
-}
-.cm_title_display{
-    float: left;
-    width: 21%;
-    font-size: 40px;
-    font-weight: 600;
-    margin-left: 6%;
-    background-image: linear-gradient(to top, #0ba360 0%, #3cba92 100%);
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent; 
-    -moz-background-clip: text;
-    -moz-text-fill-color: transparent;
-}
-.cm_title_line{
-    float: left;
-    width: 70%;
-    height: 2px;
-    margin-top: 25px;
-    background-color: rgb(44, 43, 43);
-    /* background-image: linear-gradient(to top, #6a85b6 0%, #bac8e0 100%); */
-}
-.aboutUs_row{
-    width: 100%;
-    margin-top: 50px;
-    float: left;
-}
-.coreMemberCard{
-    width: 290px;
-    height: 330px; 
-    float: left;
-    background-image: linear-gradient(rgba(0, 0, 0, 0.7), rgba(0, 0, 0, 0.7)),url(../img/swirlLiq.png);
-    background-position: center;
-    background-size:cover;
-    background-repeat: no-repeat;
-    border-radius: 3px;
-    margin-left: 6%;
-    text-align: center;
-    padding: 5px;
-    margin-bottom: 4rem;
-    color: white;
-}
-.dpcm{
-    width: 200px;
-    height: 200px;
-    border-radius: 50%;
-    box-shadow:inset 12px 12px 16px 0 rgba(0,0,0,0.25),
-       inset -8px -8px 12px 0 rgba(255,255,255,0.15);
-    padding: 5px;
-}
-.namecm{
-    text-align: center;
-    font-size: 24px;
-    font-weight: 700;
-    padding-top: 5px;
-    letter-spacing: 1px;
-}
-.desgcm{
-    text-align: center;
-    font-size: 18px;
-    font-weight: 200;
-    padding-top: 10px;
-    padding-bottom: 10px;
-}
-.cm_gfglogo{
-    width: 10%!important;
-    margin-top: -5px;
-    margin-right: 5px;
-}
-.cm_social{
-     width: 100%;
-    text-align: center;
-}
-.cm_logo1{
-    width: 10%!important;
-}
-.cm_logo{
-    width: 10%!important;
-    margin-left: 7%;
-}
-#lead_title{
-    width: 100%;
-}
-.lead_title_display{
-    float: left;
-    width: 32%;
-    font-size: 40px;
-    font-weight: 600;
-    margin-left: 6%;
-    background-image: linear-gradient(to top, #0ba360 0%, #3cba92 100%);
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent; 
-    -moz-background-clip: text;
-    -moz-text-fill-color: transparent;
-}
-.lead_title_line{
-    float: left;
-    width: 59%;
-    height: 2px;
-    margin-top: 25px;
-    background-color: rgb(44, 43, 43);
-    /* background-image: linear-gradient(to top, #6a85b6 0%, #bac8e0 100%); */
-}
-@media screen and (max-width: 1000px){
-    .cm_title_display ,.lead_title_display{
-        width: 100%;
-    }
-    .cm_title_line, .lead_title_line{
-        display: none;
-    }
-    .coreMemberCard{
-        margin-left: 18%;
-    }
-}
-/*swiper*/
-.member_title_display{
-    float: left;
-    width: 98%;
-    font-size: 40px;
-    font-weight: 600;
-    margin-left: 2%;
-    background-image: linear-gradient(to top, #0ba360 0%, #3cba92 100%);
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent; 
-    -moz-background-clip: text;
-    -moz-text-fill-color: transparent;
-    padding-bottom: 10px;
-    border-bottom: 1.5px solid black;
-}
-.swiper-container {
-    width: 100%!important;
-    padding-top: 50px !important;
-  }
-  .swiper-slide {
-    width: 250px !important;
-    height: 270px !important; 
-    background-image: linear-gradient(rgba(0, 0, 0, 0.7), rgba(0, 0, 0, 0.7)),url(../img/swirlLiq.png);
-    background-position: center;
-    background-size: cover;
-    background-repeat: no-repeat;
-    border-radius: 3px;
-    margin-left: 2%;
-    text-align: center;
-    padding-top: 5px;
-    margin-bottom: 4rem;
-    color: white;
-  }
-  .name_member{
-    text-align: center;
-    font-size: 20px;
-    font-weight: 700;
-    padding-top: 5px;
-  }
-  .desg_member{
-    text-align: center;
-    font-size: 16px;
-    font-weight: 100;
-    padding-top: 10px;
-  }
-  .dp_cm_member{
-    width: 150px;
-    height: 150px;
-    border-radius: 50%;
-    box-shadow:inset 8px 8px 12px 0 rgba(0,0,0,0.25),
-       inset -4px -4px 8px 0 rgba(255,255,255,0.15);
-    padding: 5px;
-  }
