+++ src/css/footer.css
@@ -22,14 +22,9 @@
    width: 100%;
    min-height: 100vh;
    overflow-x: hidden;
-   /* width: 100vw;
-  height: 100vh; */
    display: grid;
    place-items: center;
    z-index: 1;
-   /* position: absolute;
-   top: 0;
-   left: 0; */
    background-image: url("../img/swirl-liq-mono.png");
    background-size: cover;
    background-blend-mode: overlay;
@@ -47,21 +42,9 @@
 .footer-aboutUs {
    background-color: rgb(134, 20, 58);
 }
-
-/* .set-bg {
-  background-color: rgba(0, 0, 0, 0.6);
-  backdrop-filter: blur(25px);
-  background-size: cover;
-  width: 100%;
-  height: 100%;
-  min-height: fit-content;
-  display: grid;
-  place-items: center;
-} */
 .grid-container {
    color: white;
    font-size: 3rem;
-   /* margin: 0 10%; */
    width: 100%;
    height: 100%;
 }
@@ -80,19 +63,12 @@
 }
 .wanna-talk-to-us {
    grid-area: wanna-talk-to-us;
-
    display: flex;
    align-items: center;
-
    font-size: var(--font-size-heading);
-   /* padding-top: 8vh; */
    padding-top: 100px;
    font-family: var(--heading-font);
    font-weight: 600;
-   /* font-size: clamp(50px, 4vw, 70px); */
-
-   /* background-color: #cccccc; */
-
    margin-left: 5vw;
 }
 
@@ -103,7 +79,6 @@
    flex-direction: row;
    justify-content: flex-start;
    margin-left: 5vw;
-   /* background-color: #555; */
 }
 
 .write-to-us form {
@@ -182,12 +157,10 @@
 .links {
    grid-area: links;
    margin-right: 5vw;
-   /* background-color: #777; */
 }
 .links .flex-container {
    display: flex;
    justify-content: center;
-   /* align-items: center; */
    flex-direction: row;
    flex-wrap: nowrap;
 
@@ -219,9 +192,6 @@
 .footer-info {
    grid-area: footer-info;
    margin-right: 5vw;
-
-   /* background-color: #454545; */
-
    font-size: var(--font-size-paragragh);
 }
 .description {
@@ -237,6 +207,7 @@
    text-align: right;
    font-family: var(--paragragh-font);
    line-height: 2.2ch;
+   width: 230px;
 }
 .description2 {
    float: right;
@@ -244,6 +215,7 @@
    text-align: right;
    font-family: var(--paragragh-font);
    line-height: 2.2ch;
+   width: 280px;
 }
 .description3 {
    float: right;
@@ -251,9 +223,11 @@
    text-align: right;
    font-family: var(--paragragh-font);
    line-height: 2.2ch;
+   width: 280px;
 }
 .strong {
    font-weight: 700;
+   font-size: 18px;
    letter-spacing: 1pt;
 }
 /* Extra small devices (phones, 600px and down) */
@@ -282,6 +256,31 @@
    #email,
    #message {
       margin: 0.5vw 0;
+   }
+
+   .description1 {
+      float: right;
+      margin-right: 2vw;
+      margin-left: 0;
+      text-align: right;
+      font-family: var(--paragragh-font);
+      line-height: 2.2ch;
+      width: 200px;
+   }
+   .description2 {
+      float: right;
+      margin-right: 2vw;
+      text-align: right;
+      font-family: var(--paragragh-font);
+      line-height: 2.2ch;
+   }
+   .description3 {
+      float: right;
+      margin-right: 2vw;
+      text-align: right;
+      font-family: var(--paragragh-font);
+      line-height: 2.2ch;
+      width: 100px;
    }
 }
 @media only screen and (max-width: 800px) {
