--- src/css/navbar.css
@@ -1,188 +1,0 @@
-body {
-  margin: 0px;
-  padding: 0px;
-  font-family: montserrat;
-  overflow-x: hidden;
-}
-/* for removing scroll bar */
-::-webkit-scrollbar {
-  width: 0px;
-}
-nav{
-  padding-bottom: 10px;
-  position: absolute;
-  width: 100%;
-  z-index: 200;
-}
-nav ul {
-  float: right;
-  margin-right: 20px;
-  font-family: sans-serif;
-  font-weight: bold;
-}
-
-nav ul li {
-  display: inline-block;
-  line-height: 80px;
-  margin: 0 5px;
-}
-
-nav ul li a {
-  color: black;
-  font-size: 17px;
-  padding: 7px 13px;
-  border-radius: 3px;
-  list-style: none;
-  text-decoration: none;
-}
-
-.nav_home {
-  border-bottom: 3px solid white;
-}
-
-.nav_home:hover {
-  border-bottom: 3px solid rgb(141, 248, 141);
-}
-
-.nav_event {
-  border-bottom: 3px solid white;
-}
-
-.nav_event:hover {
-  border-bottom: 3px solid rgb(240, 77, 132);
-}
-
-.nav_projects {
-  border-bottom: 3px solid white;
-}
-
-.nav_projects:hover {
-  border-bottom: 3px solid rgb(139, 139, 235);
-}
-
-.nav_about {
-  border-bottom: 3px solid white;
-}
-
-.nav_about:hover {
-  border-bottom: 3px solid #ffc107!important;
-}
-
-.nav_checkbtn {
-  color: white;
-  text-align: right;
-  height: 60px;
-  font-size: 30px;
-  line-height: 60px;
-  margin-right: 40px;
-  float: right;
-  display: none;
-  cursor: pointer;
-}
-
-#nav_check {
-  display: none;
-}
-
-#nav_gfglogo {
-  margin-top: -10px;
-  padding-top: 20px;
-  padding-left: 3%;
-  width: 80px;
-  height: 60px;
-}
-
-@media screen and (max-width: 952px) {
-  label.logo {
-      font-size: 14px;
-      padding-left: 5%;
-  }
-
-  nav ul li a {
-      font-size: 16px;
-  }
-}
-
-@media screen and (max-width: 858px) {
-  * {
-      box-sizing: border-box;
-  }
-
-  nav for minimised screen nav background {
-      background: black;
-      height: 80px;
-      width: 100%;
-  }
-
-  .fa-bars {
-      margin-bottom: 10px;
-      font-size: 35px;
-      height: 35px;
-      width: 30px;
-  }
-
-  .nav_checkbtn {
-      display: block;
-      color: black;
-      margin-right: 8%;
-  }
-
-  ul {
-      position: fixed;
-      width: 100%;
-      height: 100vh;
-      background-color: black;
-      left: -100%;
-      text-align: center;
-      transition: all 0.5s;
-  }
-
-  nav ul li {
-      display: block;
-      margin: 50px 0%;
-      line-height: 30px;
-  }
-
-  nav ul li a {
-      font-size: 20px;
-      color: white;
-  }
-
-  .nav_home {
-      color: rgb(141, 248, 141);
-      border: none;
-  }
-
-  .nav_event {
-      color: rgb(240, 77, 132);
-      border: none;
-  }
-
-  .nav_projects {
-      color: rgb(139, 139, 235);
-      border: none;
-  }
-
-  .nav_about {
-      color: #ffc107;
-      border: none;
-  }
-
-  #nav_check:checked~ul {
-      left: 0;
-      z-index: 100;
-  }
-
-  #nav_gfglogo {
-      margin-top: -15px;
-      padding-left: 3%;
-      width: 80px;
-      height: 60px;
-  }
-}
-
-@media screen and (max-width: 500px) {
-  * {
-      box-sizing: border-box;
-  }
-}
